MPOC: Several existing Metrics rewritten for a new GriMonitoringManager.

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/monitoring/sensor/DoubleSensor.java
@@ -41,7 +41,7 @@ public class DoubleSensor extends AbstractSensor {
         value--;
     }
 
-    public void set(long value) {
+    public void set(double value) {
         this.value = value;
     }
